Changed Physics Iteration Solver to 100 (way too high, tweak this later).
Offset the Grab Point of the object when it is grabbed so that the
object is no longer intersecting with the arm when it is released. This
prevents the "throwing" behavior due to the layer on the grabbed object
being set to "Default" from "Player" abruptly, causing the engine to
rapidly expel the two intersecting bodies.